+++ flink-connectors/flink-connector-hive/src/test/java/org/apache/flink/table/catalog/hive/HiveTestUtils.java
@@ -54,7 +54,6 @@ public class HiveTestUtils {
 		HiveConf hiveConf = new HiveConf();
 		hiveConf.setVar(HiveConf.ConfVars.METASTOREWAREHOUSE, TEMPORARY_FOLDER.newFolder("hive_warehouse").getAbsolutePath());
 		hiveConf.setVar(HiveConf.ConfVars.METASTORECONNECTURLKEY, warehouseUri);
-		hiveConf.setVar(HiveConf.ConfVars.METASTOREURIS, "");
 
 		return hiveConf;
 	}
